[ConstraintSystem] NFC: Move `ConstraintLocator.h` to `include/swift/Sema`
